fix: full stack application run
The current documentation and process to run the full stack application
was not working due to a few missing pieces for both the containerized
and non-containerized versions. This commit fixes the documentation and
process to successfully run the full stack application locally.

The commit introduces a docker compose dev file which starts a frontend
and backend containers in the same network in order to allow the
frontend to reach the backend rest api.